pyproject.toml
[project]
name = "swirlyswirls"
description = "Swirly swirling particles"
version = "0.0.4"
readme = "README.md"

authors = [
    { name="Michael Lamertz", email="michael.lamertz@gmail.com" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Games/Entertainment",
    "Topic :: Software Development :: Libraries :: pygame",
]

dependencies = [
    "pygame-ce",
    "pgcooldown",
    "tinyecs",
    "pygamehelpers @ git+https://github.com/dickerdackel/pygamehelpers",
]

[project.scripts]
swirly-demo = "swirlyswirls.demo:main"

[project.urls]
homepage = "https://github.com/dickerdackel/swirlyswirls"
bugtracker = "https://github.com/DickerDackel/swirlyswirls/issues"
changelog = "https://github.com/dickerdackel/swirlyswirls/ChangeLog.md"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-rA -s --cov-report term-missing -s"
testpaths = [
    "tests",
]
